backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17
websockets==13.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
